testing: refactor X509 suite by splitting into domain-specific files
